+++ lib/util/strlcpy.c
@@ -59,6 +59,6 @@ sudo_strlcpy(char *dst, const char *src, size_t dsize)
 			continue;
 	}
 
-	return(src - osrc - 1);	/* count does not include NUL */
+	return((size_t)(src - osrc) - 1);	/* count does not include NUL */
 }
 #endif /* HAVE_STRLCPY */
